one test per bug
