+++ tests/PHPStan/Analyser/nsrt/bug10528.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Bug10528;
+
+use function PHPStan\Testing\assertType;
+
+function bug10528(string $string): void {
+	$pos = strpos('*', $string);
+	assert((bool) $pos);
+
+	assertType('int<1, max>', $pos);
+
+	$sub = substr($string, 0, $pos);
+	assert($pos !== FALSE);
+	$sub = substr($string, 0, $pos);
+}
+
