+++ tests/PHPStan/Analyser/nsrt/bug7685.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace bug7685;
+
+use function PHPStan\Testing\assertType;
+
+interface Reader {
+	public function getFilePath(): string|false;
+}
+
+function bug7685(Reader $reader): void {
+	$filePath = $reader->getFilePath();
+	if (false !== (bool) $filePath) {
+		assertType('non-falsy-string', $filePath);
+	}
+}
+
